[Serve] Reject `detached` option in `serve.start()` (#39953)
Serve offers the option to disable `detached` mode, which makes all Serve actors fate share with the driver process. Since development now occurs primarily through the CLI or `serve.run()`, this option is no longer necessary. It has since been deprecated.

This change removes all the logic that handles disabling `detached` mode. It leaves the argument in `serve.start()`, but raises an error if the argument is set to `False`. In a future release, the argument will be removed altogether.

Signed-off-by: Victor <vctr.y.m@example.com>