[Core] [Clusters] Soften ValueError for TPU autodetection (#39922)
Softens the assertion for valid TPU type to a log in TPU autodetection.

The original TPU auto detection incorrectly assumed that a GCE call would not return 200, leading to an issue where ray.init may fail if it returns status code 200 and an empty string.

This should, however, fail in gcp/config.py because we want to assert and fail ASAP if given a malformed accelerator type. We migrate the ValueError there.

We also add a specific test case here to assert that it does not crash.

Related issue number
Closes #39913

---------

Signed-off-by: allenwang28 <allencwang@google.com>
Signed-off-by: Victor <vctr.y.m@example.com>