[Serve] Reject `dedicated_cpu` option (#39955)
Serve offers the option to set the Serve controller's `num_cpus` to 1 by setting `disable_cpu` to `True`. However, the Serve controller should always set `num_cpus` to 0, so it doesn't consume any logical CPU resources.

This change removes all the internal logic handling `disable_cpu` when it's `True`, and it raises and error if the user tries to set it to anything other than `True`.

Co-authored-by: Edward Oakes <ed.nmi.oakes@gmail.com>
Signed-off-by: Victor <vctr.y.m@example.com>