[data] fix test_stats (#39949)
test_stats was timing out, so this pr increases test suite size to medium, and reapplies all of the changes in original fix.

I was also seeing some flakiness with the global memory stats locally (sometimes would spill and sometimes wouldn't) while running the entire suite, so I removed them from the output through canonicalize if we weren't testing it explicitly.

This time I'll hold off marking this as unflaky until it passes on master ci.
---------

Signed-off-by: Andrew Xue <andrewxue@anyscale.com>
Signed-off-by: Victor <vctr.y.m@example.com>