[data] estimate progress bar total (#39828)
Estimates the number of blocks outputted from the # of finished tasks and # of blocks outputted. `estimated_output_blocks = original_output_blocks * (blocks_outputted / finished_tasks)`.

Before the first task ends, we stick with the original total
<img width="1725" alt="Screenshot 2023-09-25 at 8 05 26 AM" src="https://github.com/ray-project/ray/assets/39287272/a304d0a5-0440-4972-b34d-87d03dd785ed">
And after the first task finishes (and every task after it), we update the estimated total
<img width="1723" alt="Screenshot 2023-09-25 at 8 05 34 AM" src="https://github.com/ray-project/ray/assets/39287272/0d4a0f40-f90b-4e65-a964-e8e4c3be1b46">

---------

Signed-off-by: Andrew Xue <andrewxue@anyscale.com>
Signed-off-by: Victor <vctr.y.m@example.com>