Fix error propagation for as_directory if to_directory fails (#40025)
`Checkpoint.as_directory` currently downloads the checkpoint to the local filesystem if the checkpoint is remote. It uses `to_directory` to do so. If this fails, `as_directory` will raise a strange error during its `finally` clause that is unrelated to the actual error that occurred during downloading. This PR fixes the issue by only wrapping the user's code in the `try` block. This way, any exceptions during downloading get raised immediately.

Signed-off-by: Justin Yu <justinvyu@anyscale.com>
Signed-off-by: Victor <vctr.y.m@example.com>