[Data] Propagate driver `DataContext` to `RayTrainWorkers` (#39698)
Changes made to the driver's `DataContext` are not propagated to `RayTrainWorker`s, which can lead to unexpected and inconsistent behavior between driver and workers. This PR adds the driver's `DataContext` as an initialization parameter into `RayTrainWorker`s created in a `WorkerGroup`, so that each worker will have the driver's `DataContext`.

Signed-off-by: Scott Lee <sjl@anyscale.com>
Signed-off-by: Victor <vctr.y.m@example.com>