Fix Overview Page UI will crash Issue (#40063)
Fix #39564, for an unknown reason, the backend router 'cluster_status' may return empty data without 'cluster_status'. However, the frontend may need 'cluster_status' to parse information and generate components, potentially leading to a UI crash.

The quick fix in the UI is to add protection for splitting 'cluster_status'.

We are aiming to address this issue and merge it in Ray 2.7.1.