more rllib (#39932)
Migrate rllib pytorch & tf2-static-graph learning tests to civ2. I need to declare a few tests as enormous for better test balancing. Theoretically, they are in the large categories, but many medium rllib tests are inflated and declared themselves as large (for some valid reasons), so the typical large tests need to be enourmous.

Signed-off-by: can <can@anyscale.com>