[core] rewrite the bounded_threads test with thread names and counts (#39608)
Previously we fixed an issue that when many workers send request to an async actor, it spawns threads per worker. After fixing it we have this test to assert that many workers' requests does not grow number of threads in an actor. However the test is not waterproof: even if Ray does not spawn threads, we can have threads from grpc and jemalloc.

This PR rewrites the tests to exclude "known threads" by their names, and then assert the thread counts do not grow. And if the assertion fails we print out the offending thread names to make debug easier.

Marks the test as non-flaky.