Serve Java API Improvement (#38961)
This is about Serve Java Improvement. The goal of this PR is to make the Java API consistent with Python.

The design of the user API is discussed in this proposal: ray-project/enhancements#42.
This PR only covers the Java programming API part, including Deployment.bind, Serve.run, DeploymentHandle, and so on, ensuring that these APIs can be used properly. The alignment of some internal core logic, documentation enhancements, and support for the config file will be submitted in subsequent PRs.

Signed-off-by: chuhan.ly <chuhan.ly@antgroup.com>
Co-authored-by: chuhan.ly <chuhan.ly@antgroup.com>