[Core] Increase the dashboard agent gRPC payload to 100MB & batch the metrics to multiple gRPC requests (#40177)
It seems like if there are lots of metrics to be reported, the worker fails to report metrics because the default grpc server payload size is 4MB.

This PR fixes the issue using 2 ways

Increase the default gprc server payload limit to 20MB. The Ray's default is 512MB, so I think increasing this a little bit is not risky.
When there are lots of metrics, the metrics are reported in multiple gRPC requests. It seems like each metric is about 500~1KB bytes, so we set 1000 as the default batch size (< 1MB). Note that if the default is too small, it can cause issues because agent is not that great at handling lots of GRPC requests.

Signed-off-by: Alexey Kudinkin <ak@anyscale.com>