[serve] Fix flaky test_autoscaling_policy on windows (#40411)
Originally we check that it takes at most 2 seconds for a deployment with 0 replicas to serve its first request.
This seems to be consistently enough time on Linux, but Windows is slower and often takes slightly more than 2 seconds to serve that first request. Increasing to 3 seconds to deflake.

Note that the purpose of the test is to make sure the cold start optimization works consistently, so increasing from 2 -> 3 seconds shouldn't make a difference. Without the cold start optimization, the time to serve the first request should be uniformly distributed between roughly 0-10 + cold-start-offset seconds.

Signed-off-by: Cindy Zhang <cindyzyx9@gmail.com>