[data] link dataset ids in constructor, return correct metrics id for materialize (#40413)
The dataset uuid that we record metrics with and the dataset uuid we return from `Dataset.materialize()` are different currently for two reasons:

`ExecutionPlan` is not getting it's dataset uuid as a parameter. The`dataset_uuid` parameter is only set in the deepcopy for `Dataset.serialize_lineage`. It's not useful anywhere else because we need to create the `ExecutionPlan` before the `Dataset`, and pass the plan into the dataset's constructor. Instead, we can link the dataset ids in the dataset's constructor. 

In `materialize`, we create a copy of the current dataset and execute it (this is the uuid that is used for metrics). We then create an output dataset (different uuid) and return that. This pr copies over the uuid of `copy` to `output`. Only `output` gets returned, and `copy` gets destroyed so this shouldn't create 2 datasets with same id. 

---------

Signed-off-by: Andrew Xue <andrewxue@anyscale.com>
Signed-off-by: Andrew Xue <andewzxue@gmail.com>