[core] Fix placement group invariant of PlacementResources being superset of Resources (#40038)
Looking at the code, we use RequiredResources and RequiredPlacementResources in various parts of the codebase for scheduling. This is generally fine, but since AllocateTaskResources in LocalTaskManager uses RequiredResources to allocate, and spillback/scheduling (GetBestSchedulableNode) uses RequiredPlacementResources, there is a potential infinite loop if we don't maintain the invariant of PlacementResources being a superset of Resources.

Fix it so that we do. Note the surface area of the bug was actually quite small because this invariant is maintained when num_cpus=1.

For testing:

This is already tested by test_scheduling.py in #39946