[data] add dataset name (#40430)
Adds a name field to `Dataset`. The metrics tag for a dataset is now: name + uuid.

Right now the name isn't actually stored in `Dataset` since we have no use for it here currently. This might change if we decide to use this name for other purposes.

`_dataset_name` is stored in `ExecutionPlan`. It is copied over in `copy` and `deepcopy`, so the name field of a plan stays when an operator is added to the dataset. 
---------

Signed-off-by: Andrew Xue <andewzxue@gmail.com>