Increase timeout for test basic 4 (#40492)
When we migrated the minimal test run to civ2, we run all minimal tests with two flavors: with and without redis. However, the original civ1 only run redis tests for a subset of minimal tests: https://github.com/ray-project/ray/blob/master/ci/ci.sh#L669

Fix the behavior of civ2 to align with civ1. This also fixes the flakiness from test_basic_4.py (my guess is the additional tests in civ2, which are not skipped, leaks a state that cause the flakiness here).

Signed-off-by: can <can@anyscale.com>