[serve] Lazily construct `asyncio.Event` to avoid attaching to the wrong loop (#40629)
Python 3.10 removed the `loop` parameter to `asyncio.Event`. Now, the `asyncio.Event` will call `get_running_loop` in its constructor to determine the loop to attach to. This class can be constructed for the handle from a different loop than it uses for scheduling, so we need to construct it lazily to avoid an error due to the event being attached to the wrong loop.

Signed-off-by: Edward Oakes <ed.nmi.oakes@gmail.com>