[core] Implements Session pool for the runtime env agent client. (#40579)
Previously we have a infinite thread pool in the runtime env agent client: every time a call is needed, we create a new session with new TCP port to talk to the agent. This means on high load we can run out of the ports. This PR implements a simple session pool to limit the concurrency and queue up the overflowing requests.

Also, we had a test_many_runtime_envs release test that creates lots of working_dirs and run tasks. However it did not work because each time we updated files in the dir and use it as working_dir, but the path stays the same so Ray thinks it's the same runtime env. Changed it to test on env vars.

Fixes #38662.