[Cluster launcher] [vSphere] is_terminated should consider the case that the VM is powered off to plug the GPU (#40737)
Before fix, occasionally creating worker node will fail because the worker is detected being powered off. Howwever, the power off is expected because we are plugging the GPU.

After fix, if the node has tag "vsphere-node-status:creating" and power off, we will regard it as a non-terminated node. This time, The worker node creation and join can success 100%.

Signed-off-by: Chen Jing <jingch@vmware.com>