[release][train] Avoid storage path collisions in `train_multinode_persistence` release test (#40760)
Each run of this release test now uses a unique storage path prefix, and each subtest within this uses an independent directory. Cleans up on success and leaves the logs around if it failed. Multiple runs of the release test will not conflict with each other anymore -- resolving the flakiness.

Signed-off-by: Justin Yu <justinvyu@anyscale.com>