[data] cast metadata size bytes to int (#40819)
`DefaultFileMetadataProvider` passes `size_bytes` as `np.int64` instead of python `int`. This causes a problem while reporting metrics because we require them to be either int or float. This PR casts the problematic metadata provider to provide an `int`, and also casts in `BlockMetadata` to be safe in case there are any other bad cases.

Signed-off-by: Andrew Xue <andewzxue@gmail.com>