[core] Fix the lifetime issue in Plasma server client releasing object. #40809
When the PlasmaStore wants to remove a object reference from a client in PlasmaStore::RemoveFromClientObjectIds, it somehow uses the ObjectID from the iterator from the underlying hash map. In the function MarkObjectAsUnused the hash map (object_ids) is invalidated, causing the seemingly const & object ID to be cleared out, and hence the subsequent lookup (object_ids_to_fallback_allocated_fds_.find()) always returns "not found" and the mmaps are leaked. Further, when the same object ID is mapped again, a RAY_CHECK crashes Raylet because there should be no existing mmap but we have the leaked one. This issue should have been caught by the borrow checker, but we don't have Rust yet.

This PR fixes by not using the object ID from the iterator. It does not make much sense anyway.

Somehow, the failure in python/ray/tests/test_plasma_unlimited.py only manifests in the mac os but not in Linux. I guess they have a subtle different hash map so even though the iterator is invalidated in Linux you can still access it for a while.

Signed-off-by: Ruiyang Wang <rywang014@gmail.com>