[data] Update default shuffle block size to 1GB (#40839)
With #40248, block sizes are now respected. This increases the default shuffle block size to 1GiB, which restores the previous behavior in the release test dataset_shuffle_sort_1tb. There is a possibility that this increases worker heap memory pressure during shuffle operations, but it can be resolved by overriding DataContext.

Related issue number

Closes #38400.

---------

Signed-off-by: Stephanie Wang <swang@cs.berkeley.edu>