[kuberay][autoscaler] Update KubeRay version to v1.0.0 (#40918)
Create a YAML file ray-cluster.autoscaler-template.yaml for testing instead of using the YAML file ray-cluster.autoscaler.yaml in the KubeRay repository.

Note that the tests assume that both head and worker Pods have exactly 1 CPU. Hence, if we set num-cpus: "0" in the head's rayStartParams, the current test logic would not work.
Why do I remove the test for "Confirming that the operator and autoscaler ignore pods marked for termination"?

KubeRay tries to ensure the number of runningPods, but the definition of runningPods is a bit different from different KubeRay versions.
Definition 1: For KubeRay v0.6.0 and older, the definition of runningPods is the Pods that are running or pending and not terminating.
Definition 2: For KubeRay v1.0.0, the definition of runningPods becomes that Pods that their Ray containers are not actually terminated. See [GCS FT] Consider the case of sidecar containers kuberay#1386 for more details.
That is, in definition 1, KubeRay may create new Pods if some Pods are in the terminating process. Hence, it is possible to have more than maxReplicas Pods & Ray nodes from both Kubernetes and Ray perspectives. In definition 2, KubeRay only creates new Pods when the Ray nodes are actually terminated.

---------

Signed-off-by: kaihsun <kaihsun@anyscale.com>