[serve] Flip default for `RAY_SERVE_ENABLE_NEW_HANDLE_API` to `1` (#40625)
Changes the default for the handle API to use DeploymentHandle. This will be a breaking API change. A warning was logged to inform users since Ray 2.7 (in addition to a warning on the documentation) and this will land in Ray 2.9.

Users can continue to use the existing handle API by setting the environment variable RAY_SERVE_ENABLE_NEW_HANDLE_API=0 or using handle.options(use_new_handle_api=False).

Signed-off-by: Edward Oakes <ed.nmi.oakes@gmail.com>