[Cluster launcher] [vSphere] Make sdk clients singleton, and add auto connection recovery (#40884)
In this change I did some code refactoring. In summary, the purpose are:

Make the vSphere SDK client and the pyvmomi SDK client to be singleton instances.
Add some mechanism to ensure that once the long-running session is disconnected, we can reconnect automatically.
In specific:

I added a decorator for the 2 SDKs, to make sure only one instance exists,
The connectivity check will be done every 5 mins by default for the next API to be called. If unauthorized exception is thrown, we will call the ensure_connect function of the SDK providers to do reconnect.
In the places we use the SDK, I make sure we call the init function of the clients. But acutally because of my decorator, no new instances of the client will be created. Instread of that, the decorator will help to check the connectivity, based on if the last checked time is over the constant TTL or not. We add the TTL because we don't want to check the connectivity for every original API call, that will brings too much overhead.
I removed the factory, because usually when using a factory, you usually produce a lot of instantces. In our case the instances are singleton so no need to involve the factory mode.
I added a new UT case to cover the decorator.

Signed-off-by: Chen Jing <jingch@vmware.com>
Co-authored-by: Archit Kulkarni <architkulkarni@users.noreply.github.com>