[ci] run flaky core tests serially on postmerge (#41016)
We want to test out an experiment to see if run the existing flaky tests serially will make them much more reliable. But doing so will make the job more expensive to run. I chat with @rkooo567 and we make a decision to run them serially on post-merge only as an experiement.

Signed-off-by: can <can@anyscale.com>