[Cluster launcher] [vSphere] Refactor and optimize the code (#41043)
This change is for code refactoring, including several changes:

Remove the logic for generating the keys, which has been handled by Packer
Remove one meaningless, e.g., mock-too-much, UT case. We have e2e pipline to cover that case.
Rewrite the scheduler for scheduling the next Ray node. The previous "factory mode" looks not very good.
Remove the useless code for reusing the node. On vSphere, thanks to instant clone, we don't need to cache the VM. We can always launch a new Ray node quickly. For the old ones, we just delete the VM after the Ray process on it quits the Ray cluster.
Remove the cache for the nodes. The cache could be outdated when the node information changes on vSphere side. We have only one place leverages the node cache so the benefit for keeping the cache doesn't worth its potential issue.
Moved some functions which have no dependencies on the VsphereNodeProvider instance to the outside scope, instead of being a member function of the VsphereNodeProvider instance.
Some changes caused by above changes.
Some Python grammar level optimization.

Signed-off-by: Chen Jing <jingch@vmware.com>