[core] Add MPI support on Ray cluster (#40917)
This PR adds the support to run MPI based code on top of Ray.

The support is done with runtime env plugin. To enable it, the following decorator needs to be added inside ray remote options:

@ray.remote(
        runtime_env={
            "mpi": {
                "args": ["-n", "4"],
                "worker_entry": "mpi_worker.run",
            }
        }
)
def f():
    pass
Here the mpi_worker.run is the function the process with rank > 0 will run. It'll run as import mpi_worker; mpi_worker.run(). The parameter needs to be passed with MPI comm.bcast.

Here the process with rank 0 sill will run the remote function f.

Signed-off-by: Yi Cheng <74173148+iycheng@users.noreply.github.com>