[Data] Fix `FileBasedDatasource` serialization (#41061)
When you launch a read task generated byFileBasedDatasource, Ray serializes the FileBasedDatasource instance because the read task calls FileBasedDatasource._read_stream. This wasn't an issue before, because FileBasedDatasource was stateless and therefore quick to serialize. However, #40900 moved state like the input file paths from _FileBasedDatasourceReader to FileBasedDatasource. As a result, FileBasedDatasource is now slow to serialize and read tasks can be slow to launch.

To fix this issue, this PR stores paths and file sizes in the object store, as only stores references to them in FileBasedDatasource.

---------

Signed-off-by: Balaji Veeramani <balaji@anyscale.com>