[CI] [GCP Cluster launcher] Use single SSH key for all tests (#40677)
Previously a new key would be generated by the cluster launcher for each test run. This caused us to hit the limit for the number of SSH keys per project. This PR uses a single SSH key for all tests by storing a key in a GCS bucket and downloading it to each test cluster at startup.

Related issue number
Closes #40635

---------

Signed-off-by: Archit Kulkarni <architkulkarni@users.noreply.github.com>
Signed-off-by: Archit Kulkarni <archit@anyscale.com>
Signed-off-by: Xiaoshuang Liu <lxsbit@126.com>