[image] directly use rm rather than mv in ray-prestart (#334)
moving a big directory on overlay fs turns out to be very slow. more
than half of the prestart time ends up used for just moving the files.

so just remove the files. in case the script fails, we saved another
copy of the OSS ray tarball under /opt/anyscale, which can be placed
back