[Cluster launcher] [vSphere] dynamic passthrough gpu support (#41087)
From ESXI 7.0’s release there was an addition called Dynamic DirectPath IO which is an option for GPU passthrough.

DirectPath IO is a feature that allows a physical PCIe device to be directly mapped to a VM, similar to Dynamic DirectPath IO. The main difference between the two is that DirectPath IO allows the VM to have direct access to the physical device all the time, while Dynamic DirectPath IO allows the VM to access the physical device only when the VM is powered on.

This is to support Dynamic DirectPath IO. If you want to enable, do it like this:

provider:
    ...
    vsphere_config:
      .....
      gpu_config:
        dynamic_pci_passthrough: True

Signed-off-by: Chen Hui <huchen@vmware.com>
Signed-off-by: Xiaoshuang Liu <lxsbit@126.com>