[data] fix split read output blocks (#41070)
I'm not sure why we were calling `_set_num_output_blocks` on the upstream input operator when we were splitting blocks for the downstream read map operator. The split happens in the read map operator, which shouldn't affect the input operator.

Also adds a test for estimating output blocks with additional split factor.
---------

Signed-off-by: Andrew Xue <andewzxue@gmail.com>
Signed-off-by: Xiaoshuang Liu <lxsbit@126.com>