[Serve] Fix batch disconnect issue (#41199)
In batching requests, after the client dropped the connection and the future is canceled, we might still be calling set_result on the future object and cause the entire batching processing loop to terminate. This PR fixes it by only calling set_result on not canceled futures. Also, added tests for both unary and streaming cases.

Related issue number
Closes #41057

---------

Signed-off-by: Gene Su <e870252314@gmail.com>
Signed-off-by: Gene Der Su <gdsu@ucdavis.edu>
Co-authored-by: shrekris-anyscale <92341594+shrekris-anyscale@users.noreply.github.com>
Signed-off-by: Xiaoshuang Liu <lxsbit@126.com>